docs: add comments

+++ api/paywall/database.js
@@ -1,6 +1,7 @@
 import { sequelize } from '../../shared/server';
 import * as Sequelize from 'sequelize';
 
+// Defines model for article view
 export const ArticleView = sequelize.define('article_view', {
   visitorId: {
     type: Sequelize.STRING,
@@ -13,6 +14,14 @@ export const ArticleView = sequelize.define('article_view', {
   },
 });
 
+let didInit = false;
+
 export async function initPaywall() {
+  if (didInit) {
+    return;
+  }
+
   await ArticleView.sync({ force: false });
+
+  didInit = true;
 }
